+++ commands/setup.js
@@ -57,6 +57,6 @@ exports.conf = {
 exports.help = {
     name: "setup",
     description: "Starts a game setup menu, allowing you to invite other players or set game options. Once setup has started, you can also use this command to invite players and set up game options.",
-    usage: "setup [game] OR setupgame [invite/option/turns/resend/leave/kick/start/cancel] [...]",
+    usage: "setup [game] OR setupgame [invite/option/turns/variant/resend/leave/kick/start/cancel] [...]",
     examples: ["setup Tic-tac-toe", "setup invite @User#1234 @User2#5678", "setup turns 1 @User#1234"]
 };